tests/wpt/web-platform-tests/speculation-rules/prefetch/same-origin-cookies.https.html

<!DOCTYPE html>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src='/resources/testdriver-vendor.js'></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="/common/utils.js"></script>
<script src="resources/utils.sub.js"></script>
<script>
  promise_test(async t => {
    assert_implements(HTMLScriptElement.supports('speculationrules'), "Speculation Rules not supported");

    await test_driver.delete_all_cookies();

    let executor = 'cookies.py';
    let agent = await spawnWindow(t, { executor });
    let response_cookies = await agent.getResponseCookies();
    let request_cookies = await agent.getRequestCookies();
    assert_equals(request_cookies["count"], undefined);
    assert_equals(request_cookies["type"], undefined);
    assert_equals(response_cookies["count"], "1");
    assert_equals(response_cookies["type"], "navigate");

    let nextUrl = agent.getExecutorURL({ executor, page: 2 });
    await agent.forceSinglePrefetch(nextUrl);
    await agent.navigate(nextUrl);

    response_cookies = await agent.getResponseCookies();
    request_cookies = await agent.getRequestCookies();
    assert_equals(request_cookies["count"], "1");
    assert_equals(request_cookies["type"], "navigate");
    assert_equals(response_cookies["count"], "2");
    assert_equals(response_cookies["type"], "prefetch");

    assert_prefetched(await agent.getRequestHeaders());

  }, "speculation rules based prefetch should use cookies for same origin urls.");
</script>
